fix(tasks): remove private composition edge
