+++ .ci/behavior-ownership/partition.v1.json
@@ -29,10 +29,6 @@
       "target": "backend/app/adapters/project_agents/openai_agent_sdk.py"
     },
     {
-      "group": "lifecycle",
-      "target": "backend/app/adapters/tasks/submission_composition.py"
-    },
-    {
       "group": "shared",
       "target": "backend/app/api/deps/api_controls.py"
     },
@@ -829,7 +825,7 @@
       "target": "backend/scripts/week2_api_e2e.py"
     }
   ],
-  "authority_digest": "4a06d3ea196d5aa9972b2c07019401f5fd5a8f44049745024359660e11a92b08",
+  "authority_digest": "6e8c9f70a0917f12e28205d7056404a29e1a0611992b7d87b59c54f96e2b5ea4",
   "protected_base_commit": "7676ce4347db0c9694962a9b587a20765e16eac6",
   "schema": "workstream.behavior-ownership-partition.v1"
 }
